+++ DocumentationTests/DocumentationTests/DocumentationTests/05-ui-cookbook/17-camera-zoom.swift
@@ -1,0 +1,12 @@
+import StreamVideo
+import StreamVideoSwiftUI
+import SwiftUI
+import Combine
+import AVFoundation
+
+@MainActor
+fileprivate func content() {
+    container {
+        try call.zoom(by: 1.5)
+    }
+}