+++ Sources/StreamVideo/WebRTC/VideoCapturer.swift
@@ -11,9 +11,9 @@ class VideoCapturer: CameraVideoCapturing {
     private var videoOptions: VideoOptions
     private let videoSource: RTCVideoSource
     private var videoCaptureHandler: StreamVideoCaptureHandler?
-
+    
     private var simulatorStreamFile: URL? = InjectedValues[\.simulatorStreamFile]
-
+    
     init(
         videoSource: RTCVideoSource,
         videoOptions: VideoOptions,
@@ -104,7 +104,7 @@ class VideoCapturer: CameraVideoCapturing {
             }
         }
     }
-
+    
     /// Initiates a focus and exposure operation at the specified point on the camera's view.
     ///
     /// This method attempts to focus the camera and set the exposure at a specific point by interacting
@@ -122,43 +122,134 @@ class VideoCapturer: CameraVideoCapturing {
     /// from (0,0) at the top-left to (1,1) at the bottom-right.
     func focus(at point: CGPoint) throws {
         guard
-            let captureSession = (videoCapturer as? RTCCameraVideoCapturer)?.captureSession,
-            let device = captureSession.inputs.first as? AVCaptureDeviceInput
+            let cameraVideoCapturer = videoCapturer as? RTCCameraVideoCapturer,
+            let activeCaptureDevice = cameraVideoCapturer.captureSession.activeVideoCaptureDevice
         else {
             throw ClientError.Unexpected()
         }
-
-        do {
-            try device.device.lockForConfiguration()
-
-            if device.device.isFocusPointOfInterestSupported {
-                log.debug("Will focus at point: \(point)")
-                device.device.focusPointOfInterest = point
-
-                if device.device.isFocusModeSupported(.autoFocus) {
-                    device.device.focusMode = .autoFocus
+        
+        try activeCaptureDevice.lockForConfiguration()
+        
+        if activeCaptureDevice.isFocusPointOfInterestSupported {
+            log.debug("Will focus at point: \(point)")
+            activeCaptureDevice.focusPointOfInterest = point
+            
+            if activeCaptureDevice.isFocusModeSupported(.autoFocus) {
+                activeCaptureDevice.focusMode = .autoFocus
+            } else {
+                log.warning("There are no supported focusMode.")
+            }
+            
+            log.debug("Will set exposure at point: \(point)")
+            if activeCaptureDevice.isExposurePointOfInterestSupported {
+                activeCaptureDevice.exposurePointOfInterest = point
+                
+                if activeCaptureDevice.isExposureModeSupported(.autoExpose) {
+                    activeCaptureDevice.exposureMode = .autoExpose
                 } else {
-                    log.warning("There are no supported focusMode.")
-                }
-
-                log.debug("Will set exposure at point: \(point)")
-                if device.device.isExposurePointOfInterestSupported {
-                    device.device.exposurePointOfInterest = point
-
-                    if device.device.isExposureModeSupported(.autoExpose) {
-                        device.device.exposureMode = .autoExpose
-                    } else {
-                        log.warning("There are no supported exposureMode.")
-                    }
+                    log.warning("There are no supported exposureMode.")
                 }
             }
-
-            device.device.unlockForConfiguration()
-        } catch {
-            log.error(error)
         }
+        
+        activeCaptureDevice.unlockForConfiguration()
     }
-
+    
+    /// Adds the `AVCapturePhotoOutput` on the `CameraVideoCapturer` to enable photo
+    /// capturing capabilities.
+    ///
+    /// This method configures the local user's `CameraVideoCapturer` with an
+    /// `AVCapturePhotoOutput` for capturing photos. This enhancement allows applications to capture
+    /// still images while video capturing is ongoing.
+    ///
+    /// - Parameter capturePhotoOutput: The `AVCapturePhotoOutput` instance to be added
+    /// to the `CameraVideoCapturer`. This output enables the capture of photos alongside video
+    /// capturing.
+    ///
+    /// - Throws: An error if the `CameraVideoCapturer` does not support adding an `AVCapturePhotoOutput`.
+    /// This method is specifically designed for `RTCCameraVideoCapturer` instances. If the
+    /// `CameraVideoCapturer` in use does not support photo output functionality, an appropriate error
+    /// will be thrown to indicate that the operation is not supported.
+    ///
+    /// - Warning: A maximum of one output of each type may be added.
+    func addCapturePhotoOutput(_ capturePhotoOutput: AVCapturePhotoOutput) throws {
+        guard
+            let cameraVideoCapturer = videoCapturer as? RTCCameraVideoCapturer,
+            cameraVideoCapturer.captureSession.canAddOutput(capturePhotoOutput)
+        else {
+            throw ClientError.Unexpected("Cannot set capturePhotoOutput for videoCapturer of type:\(type(of: videoCapturer)).")
+        }
+        
+        cameraVideoCapturer.captureSession.beginConfiguration()
+        cameraVideoCapturer.captureSession.addOutput(capturePhotoOutput)
+        cameraVideoCapturer.captureSession.commitConfiguration()
+    }
+    
+    /// Adds an `AVCaptureVideoDataOutput` to the `CameraVideoCapturer` for video frame
+    /// processing capabilities.
+    ///
+    /// This method configures the local user's `CameraVideoCapturer` with an
+    /// `AVCaptureVideoDataOutput`, enabling the processing of video frames. This is particularly
+    /// useful for applications that require access to raw video data for analysis, filtering, or other processing
+    /// tasks while video capturing is in progress.
+    ///
+    /// - Parameter videoOutput: The `AVCaptureVideoDataOutput` instance to be added to
+    /// the `CameraVideoCapturer`. This output facilitates the capture and processing of live video
+    /// frames.
+    ///
+    /// - Throws: An error if the `CameraVideoCapturer` does not support adding an
+    /// `AVCaptureVideoDataOutput`. This functionality is specific to `RTCCameraVideoCapturer`
+    /// instances. If the current `CameraVideoCapturer` does not accommodate video output, an error
+    /// will be thrown to signify the unsupported operation.
+    ///
+    /// - Warning: A maximum of one output of each type may be added. For applications linked on or
+    /// after iOS 16.0, this restriction no longer applies to AVCaptureVideoDataOutputs. When adding more
+    /// than one AVCaptureVideoDataOutput, AVCaptureSession.hardwareCost must be taken into account.
+    func addVideoOutput(_ videoOutput: AVCaptureVideoDataOutput) throws {
+        guard
+            let cameraVideoCapturer = videoCapturer as? RTCCameraVideoCapturer,
+            cameraVideoCapturer.captureSession.canAddOutput(videoOutput)
+        else {
+            throw ClientError.Unexpected("Cannot set videoOutput for videoCapturer of type:\(type(of: videoCapturer)).")
+        }
+        cameraVideoCapturer.captureSession.beginConfiguration()
+        cameraVideoCapturer.captureSession.addOutput(videoOutput)
+        cameraVideoCapturer.captureSession.commitConfiguration()
+    }
+    
+    /// Zooms the camera video by the specified factor.
+    ///
+    /// This method attempts to zoom the camera's video feed by adjusting the `videoZoomFactor` of
+    /// the camera's active device. It first checks if the video capturer is of type `RTCCameraVideoCapturer`
+    /// and if the current camera device supports zoom by verifying that the `videoMaxZoomFactor` of
+    /// the active format is greater than 1.0. If these conditions are met, it proceeds to apply the requested
+    /// zoom factor, clamping it within the supported range to avoid exceeding the device's capabilities.
+    ///
+    /// - Parameter factor: The desired zoom factor. A value of 1.0 represents no zoom, while values
+    /// greater than 1.0 increase the zoom level. The factor is clamped to the maximum zoom factor supported
+    /// by the device to ensure it remains within valid bounds.
+    ///
+    /// - Throws: `ClientError.Unexpected` if the video capturer is not of type
+    /// `RTCCameraVideoCapturer`, or if the device does not support zoom. Also, throws an error if
+    /// locking the device for configuration fails.
+    ///
+    /// - Note: This method should be used cautiously, as setting a zoom factor significantly beyond the
+    /// optimal range can degrade video quality.
+    func zoom(by factor: CGFloat) throws {
+        guard
+            let cameraVideoCapturer = videoCapturer as? RTCCameraVideoCapturer,
+            let activeCaptureDevice = cameraVideoCapturer.captureSession.activeVideoCaptureDevice,
+            activeCaptureDevice.activeFormat.videoMaxZoomFactor > 1.0 // That ensures that the devices supports zoom.
+        else {
+            throw ClientError.Unexpected("Cannot zoom captureDevice for videoCapturer of type:\(type(of: videoCapturer)).")
+        }
+        
+        try activeCaptureDevice.lockForConfiguration()
+        let zoomFactor = max(1.0, min(factor, activeCaptureDevice.activeFormat.videoMaxZoomFactor))
+        activeCaptureDevice.videoZoomFactor = zoomFactor
+        activeCaptureDevice.unlockForConfiguration()
+    }
+    
     // MARK: - private
     
     private func checkForBackgroundCameraAccess() {
@@ -168,7 +259,7 @@ class VideoCapturer: CameraVideoCapturing {
             }
             // Configure the capture session.
             captureSession.beginConfiguration()
-
+            
             if captureSession.isMultitaskingCameraAccessSupported {
                 // Enable use of the camera in multitasking modes.
                 captureSession.isMultitaskingCameraAccessEnabled = true
@@ -190,7 +281,7 @@ extension CMVideoDimensions {
 }
 
 extension AVCaptureDevice.Format {
-
+    
     // computes a ClosedRange of supported FPSs for this format
     func fpsRange() -> ClosedRange<Int> {
         videoSupportedFrameRateRanges
@@ -202,7 +293,7 @@ extension AVCaptureDevice.Format {
 }
 
 extension AVFrameRateRange {
-
+    
     // convert to a ClosedRange
     func toRange() -> ClosedRange<Int> {
         Int(minFrameRate)...Int(maxFrameRate)
@@ -217,7 +308,7 @@ internal func merge<T>(
 }
 
 extension Comparable {
-
+    
     // clamp a value within the range
     func clamped(to limits: ClosedRange<Self>) -> Self {
         min(max(self, limits.lowerBound), limits.upperBound)